+++ LifeCycleAssessment_Engine/LifeCycleAssessment_Engine.csproj
@@ -140,7 +140,7 @@
     <Compile Include="Compute\EvaluateEnvironmentalProductDeclarationByMass.cs" />
     <Compile Include="Compute\EvaluateLifeCycleAssessment.cs" />
     <Compile Include="Compute\EvaluateLifeCycleAssessmentScope.cs" />
-    <Compile Include="Modify\Collapse.cs" />
+    <Compile Include="Compute\Collapse.cs" />
     <Compile Include="Query\GetQuantityType.cs" />
     <Compile Include="Query\GetQuantityTypeValue.cs" />
     <Compile Include="Query\MaterialEndOfLifeTreatment.cs" />
@@ -160,6 +160,9 @@
     <Compile Include="Query\EnvironmentalProductDeclaration.cs" />
     <Compile Include="Query\GetEvaluationValue.cs" />
     <Compile Include="Query\TotalGlobalWarmingPotential.cs" />
+  </ItemGroup>
+  <ItemGroup>
+    <Folder Include="Modify\" />
   </ItemGroup>
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
   <PropertyGroup>
